refactor : Rank, File을 Position으로부터 분리
